+++ apps/funding-service/src/db/index.spec.ts
@@ -2,9 +2,10 @@ import assert from "assert";
 import fs from "fs";
 import { IBackup, IMemoryDb, newDb } from "pg-mem";
 import * as db from ".";
-import { CreateAccessToken, generateAccessToken } from "../access-token";
+import { CreateAccessToken } from "../access-token/dto";
+import { generateAccessToken } from "../utils";
 import { DBInstance } from ".";
-import { CreateRequest, UpdateRequest } from "../request";
+import { CreateRequest, UpdateRequest } from "../request/dto";
 import { utils } from "@rpch/common";
 import path from "path";
 import * as fixtures from "@rpch/common/build/fixtures";
